backend/requirements.txt
# NutriSnap Backend Dependencies
# Python 3.10+ required

# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Database
motor==3.3.2  # Async MongoDB driver
pymongo==4.6.1

# Data Validation & Settings
pydantic==2.6.1
pydantic-settings==2.1.0

# Environment Variables
python-dotenv==1.0.1

# AI Orchestration
langchain==0.1.20
langchain-google-genai==0.0.11
langchain-groq==0.0.1

# HTTP Client (for CalorieNinjas API)
httpx==0.26.0

# File Upload Support
python-multipart==0.0.9

# Image Processing (for future enhancements)
Pillow==10.2.0

# Type Checking & Linting (development)
mypy==1.8.0
black==24.1.1
